js: Port integration tests to its own package
